wrapper update AP-754
- Added CreateDoiContact
- Added UploadImageToGallery
- Added parameters  startDate and endDate for GetContactCampaignStats
- Added parameters from and tags for GetTransacEmailsListTransactionalEmails
- Modified datatype of sms and email parameters from int to double in RemainingCreditModelReseller
- Documentation updates

+++ lib/Model/RemainingCreditModelChild.php
@@ -58,8 +58,8 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerTypes = [
-        'sms' => 'int',
-        'email' => 'int'
+        'sms' => 'double',
+        'email' => 'double'
     ];
 
     /**
@@ -68,8 +68,8 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'sms' => 'int64',
-        'email' => 'int64'
+        'sms' => 'double',
+        'email' => 'double'
     ];
 
     /**
@@ -220,7 +220,7 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
     /**
      * Gets sms
      *
-     * @return int
+     * @return double
      */
     public function getSms()
     {
@@ -230,7 +230,7 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
     /**
      * Sets sms
      *
-     * @param int $sms SMS Credits remaining for child account
+     * @param double $sms SMS Credits remaining for child account
      *
      * @return $this
      */
@@ -244,7 +244,7 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
     /**
      * Gets email
      *
-     * @return int
+     * @return double
      */
     public function getEmail()
     {
@@ -254,7 +254,7 @@ class RemainingCreditModelChild implements ModelInterface, ArrayAccess
     /**
      * Sets email
      *
-     * @param int $email Email Credits remaining for child account
+     * @param double $email Email Credits remaining for child account
      *
      * @return $this
      */
